pyproject.toml
[project]
name = "MISTRAL_FLEET_OPS"
version = "0.1.0"
description = "Mistral Fleet Ops - MCP server for parallel game deployment across Blaxel sandboxes."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.12.3",
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.1",
    "blaxel>=0.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    "ruff>=0.12.12",
    "weave>=0.52.7",
    "qdrant-client>=1.9.0",
    "groq>=0.4.0",
    "requests>=2.31.0",
    # LlamaIndex - Production RAG Framework
    "llama-index>=0.10.0",
    "llama-index-embeddings-mistralai>=0.1.0",
    "llama-index-vector-stores-qdrant>=0.2.0",
    "llama-index-llms-mistralai>=0.1.0",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "I"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
